+++ CHANGELOG.md
@@ -2,4 +2,7 @@
 
 ## 1.0.0 (Soon)
 
-### 
+### Features
+
+- added `Stack`, `Spacer` components
+- added `LayoutProvider`